Seperated Components and Files into Different Folders
