+++ frontend/src/pages/Chat/components/ChannelMembers/index.js
@@ -1,0 +1,2 @@
+// src/pages/Chat/components/ChannelMembers/index.js
+export { default } from './ChannelMembers';
