+++ frontend/src/pages/Chat/components/CustomChannelPreview/index.js
@@ -1,0 +1,2 @@
+// src/pages/Chat/components/CustomChannelPreview/index.js
+export { default } from './CustomChannelPreview';
