+++ frontend/src/pages/Chat/modals/AddMembersModal/index.js
@@ -1,0 +1,2 @@
+// src/pages/Chat/modals/AddMembersModal/index.js
+export { default } from './AddMembersModal';
