+++ frontend/src/pages/Chat/modals/ConfirmationModal/ConfirmationModal.js
@@ -1,0 +1,33 @@
+// src/pages/Chat/modals/ConfirmationModal/ConfirmationModal.js
+import React from 'react';
+import { X } from 'lucide-react';
+
+const ConfirmationModal = ({ isOpen, onClose, onConfirm, title, message }) => {
+  if (!isOpen) return null;
+
+  return (
+    <div className="modal-overlay" onClick={onClose}>
+      <div className="modal-content" onClick={(e) => e.stopPropagation()}>
+        <div className="modal-header">
+          <h3>{title}</h3>
+          <button onClick={onClose} className="modal-close">
+            <X />
+          </button>
+        </div>
+        <div className="modal-body">
+          <p>{message}</p>
+          <div className="confirmation-actions">
+            <button className="cancel-button" onClick={onClose}>
+              Cancel
+            </button>
+            <button className="confirm-button" onClick={onConfirm}>
+              Confirm
+            </button>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default ConfirmationModal;
