+++ frontend/src/pages/Chat/modals/ConfirmationModal/index.js
@@ -1,0 +1,2 @@
+// src/pages/Chat/modals/ConfirmationModal/index.js
+export { default } from './ConfirmationModal';
