+++ frontend/src/pages/Chat/modals/RemoveMembersModal/index.js
@@ -1,0 +1,2 @@
+// src/pages/Chat/modals/RemoveMembersModal/index.js
+export { default } from './RemoveMembersModal';
